refactor: register-command
