+++ src/main/kotlin/dev/m2en/citation/Main.kt
@@ -4,7 +4,6 @@ package dev.m2en.citation
 
 import dev.kord.core.Kord
 import dev.kord.core.behavior.interaction.response.respond
-import dev.kord.core.entity.ReactionEmoji
 import dev.kord.core.event.gateway.ReadyEvent
 import dev.kord.core.event.interaction.GuildChatInputCommandInteractionCreateEvent
 import dev.kord.core.event.interaction.GuildMessageCommandInteractionCreateEvent
@@ -16,8 +15,8 @@ import dev.kord.gateway.Intents
 import dev.kord.core.event.message.ReactionAddEvent
 import dev.kord.gateway.PrivilegedIntent
 import dev.m2en.citation.command.chat.HelpCommand
+import dev.m2en.citation.command.message.chat.RegisterCommand
 import dev.m2en.citation.command.message.onDebugMessageCommand
-import dev.m2en.citation.command.message.onRegister
 import dev.m2en.citation.handler.InteractionCommandInterface
 import dev.m2en.citation.message.QuoteDeleteListener
 import dev.m2en.citation.message.QuoteSendListener
@@ -31,8 +30,6 @@ suspend fun main() {
     val interactionMap = mutableMapOf<String, InteractionCommandInterface>()
     interactionMap["help"] = HelpCommand
 
-    val reactionEmoji = ReactionEmoji.Unicode("\uD83D\uDDD1")
-
     kord.on<ReadyEvent> {
         println("citation is ready!\n実行中バージョン: v" + getCitationVersion())
     }
@@ -40,7 +37,7 @@ suspend fun main() {
     kord.on<MessageCreateEvent> {
         if(message.getGuildOrNull() == null) return@on
 
-        onRegister(reactionEmoji)
+        RegisterCommand.messageHandle(message)
         QuoteSendListener.messageHandle(message)
     }
